+++ buildSrc/src/main/kotlin/Config.kt
@@ -17,7 +17,7 @@
 
 object Realm {
     val ciBuild = (System.getenv("JENKINS_HOME") != null)
-    const val version = "1.4.1-SNAPSHOT"
+    const val version = "1.5.0"
     const val group = "io.realm.kotlin"
     const val projectUrl = "https://realm.io"
     const val pluginPortalId = "io.realm.kotlin"
@@ -71,7 +71,7 @@ object Versions {
     const val cmake = "3.22.1"
     const val coroutines = "1.6.4" // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-core
     const val datetime = "0.4.0" // https://github.com/Kotlin/kotlinx-datetime
-    const val detektPlugin = "1.19.0-RC1" // https://github.com/detekt/detekt
+    const val detektPlugin = "1.22.0-RC2" // https://github.com/detekt/detekt
     const val dokka = "1.6.0" // https://github.com/Kotlin/dokka
     const val gradlePluginPublishPlugin = "0.15.0" // https://plugins.gradle.org/plugin/com.gradle.plugin-publish
     const val jmh = "1.34" // https://github.com/openjdk/jmh
@@ -89,6 +89,7 @@ object Versions {
     const val serialization = "1.4.0" // https://kotlinlang.org/docs/releases.html#release-details
     const val shadowJar =  "6.1.0" // https://mvnrepository.com/artifact/com.github.johnrengelman.shadow/com.github.johnrengelman.shadow.gradle.plugin?repo=gradle-plugins
     const val multidex = "2.0.1" // https://developer.android.com/jetpack/androidx/releases/multidex
+    const val kbson = "0.1.0" // https://github.com/mongodb/kbson
 }
 
 // Could be actual Dependency objects