+++ packages/cinterop/src/jvm/kotlin/io/realm/kotlin/internal/interop/RealmInterop.kt
@@ -18,6 +18,7 @@ package io.realm.kotlin.internal.interop
 
 import io.realm.kotlin.internal.interop.Constants.ENCRYPTION_KEY_LENGTH
 import io.realm.kotlin.internal.interop.RealmInterop.cptr
+import io.realm.kotlin.internal.interop.sync.ApiKeyWrapper
 import io.realm.kotlin.internal.interop.sync.AuthProvider
 import io.realm.kotlin.internal.interop.sync.CoreSubscriptionSetState
 import io.realm.kotlin.internal.interop.sync.CoreSyncSessionState
@@ -33,6 +34,7 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.CoroutineStart
 import kotlinx.coroutines.launch
+import org.mongodb.kbson.ObjectId
 
 // FIXME API-CLEANUP Rename io.realm.interop. to something with platform?
 //  https://github.com/realm/realm-kotlin/issues/56
@@ -81,12 +83,15 @@ actual object RealmInterop {
 
         for ((i, entry) in schema.withIndex()) {
             val (clazz, properties) = entry
+
+            val computedCount = properties.count { it.isComputed }
+
             // Class
             val cclass = realm_class_info_t().apply {
                 name = clazz.name
                 primary_key = clazz.primaryKey
-                num_properties = properties.size.toLong()
-                num_computed_properties = 0
+                num_properties = (properties.size - computedCount).toLong()
+                num_computed_properties = computedCount.toLong()
                 key = INVALID_CLASS_KEY.key
                 flags = clazz.flags
             }
@@ -162,6 +167,10 @@ actual object RealmInterop {
 
     actual fun realm_config_set_data_initialization_function(config: RealmConfigurationPointer, callback: DataInitializationCallback) {
         realmc.realm_config_set_data_initialization_function(config.cptr(), callback)
+    }
+
+    actual fun realm_config_set_in_memory(config: RealmConfigurationPointer, inMemory: Boolean) {
+        realmc.realm_config_set_in_memory(config.cptr(), inMemory)
     }
 
     actual fun realm_open(config: RealmConfigurationPointer, dispatcher: CoroutineDispatcher?): Pair<LiveRealmPointer, Boolean> {
@@ -472,6 +481,16 @@ actual object RealmInterop {
         )
     }
 
+    actual fun realm_get_backlinks(obj: RealmObjectPointer, sourceClassKey: ClassKey, sourcePropertyKey: PropertyKey): RealmResultsPointer {
+        return LongPointerWrapper(
+            realmc.realm_get_backlinks(
+                (obj as LongPointerWrapper).ptr,
+                sourceClassKey.key,
+                sourcePropertyKey.key
+            )
+        )
+    }
+
     actual fun realm_list_size(list: RealmListPointer): Long {
         val size = LongArray(1)
         realmc.realm_list_size(list.cptr(), size)
@@ -679,6 +698,7 @@ actual object RealmInterop {
     }
 
     private fun initIndicesArray(size: LongArray): LongArray = LongArray(size[0].toInt())
+    @Suppress("UnusedPrivateMember")
     private fun initRangeArray(size: LongArray): Array<LongArray> = Array(size[0].toInt()) { LongArray(2) }
 
     actual fun <T, R> realm_collection_changes_get_indices(change: RealmChangesPointer, builder: CollectionChangeSetBuilder<T, R>) {
@@ -937,6 +957,16 @@ actual object RealmInterop {
         realmc.realm_sync_client_config_set_metadata_mode(
             syncClientConfig.cptr(),
             metadataMode.nativeValue
+        )
+    }
+
+    actual fun realm_sync_client_config_set_metadata_encryption_key(
+        syncClientConfig: RealmSyncClientConfigurationPointer,
+        encryptionKey: ByteArray
+    ) {
+        realmc.realm_sync_client_config_set_metadata_encryption_key(
+            syncClientConfig.cptr(),
+            encryptionKey
         )
     }
 
@@ -1369,11 +1399,11 @@ actual object RealmInterop {
         return LongPointerWrapper(realmc.realm_flx_sync_config_new(user.cptr()))
     }
 
-    actual fun realm_sync_subscription_id(subscription: RealmSubscriptionPointer): ObjectIdWrapper {
+    actual fun realm_sync_subscription_id(subscription: RealmSubscriptionPointer): ObjectId {
         val nativeBytes: ShortArray = realmc.realm_sync_subscription_id(subscription.cptr()).bytes
         val byteArray = ByteArray(nativeBytes.size)
         nativeBytes.mapIndexed { index, b -> byteArray[index] = b.toByte() }
-        return ObjectIdWrapperImpl(byteArray)
+        return ObjectId(byteArray)
     }
 
     actual fun realm_sync_subscription_name(subscription: RealmSubscriptionPointer): String? {
@@ -1587,6 +1617,88 @@ actual object RealmInterop {
         }
     }
 
+    actual fun realm_app_user_apikey_provider_client_create_apikey(
+        app: RealmAppPointer,
+        user: RealmUserPointer,
+        name: String,
+        callback: AppCallback<ApiKeyWrapper>
+    ) {
+        realmc.realm_app_user_apikey_provider_client_create_apikey(
+            app.cptr(),
+            user.cptr(),
+            name,
+            callback
+        )
+    }
+
+    actual fun realm_app_user_apikey_provider_client_delete_apikey(
+        app: RealmAppPointer,
+        user: RealmUserPointer,
+        id: ObjectId,
+        callback: AppCallback<Unit>
+    ) {
+        realmc.realm_app_user_apikey_provider_client_delete_apikey(
+            app.cptr(),
+            user.cptr(),
+            id.asRealmObjectIdT(),
+            callback
+        )
+    }
+
+    actual fun realm_app_user_apikey_provider_client_disable_apikey(
+        app: RealmAppPointer,
+        user: RealmUserPointer,
+        id: ObjectId,
+        callback: AppCallback<Unit>
+    ) {
+        realmc.realm_app_user_apikey_provider_client_disable_apikey(
+            app.cptr(),
+            user.cptr(),
+            id.asRealmObjectIdT(),
+            callback
+        )
+    }
+
+    actual fun realm_app_user_apikey_provider_client_enable_apikey(
+        app: RealmAppPointer,
+        user: RealmUserPointer,
+        id: ObjectId,
+        callback: AppCallback<Unit>
+    ) {
+        realmc.realm_app_user_apikey_provider_client_enable_apikey(
+            app.cptr(),
+            user.cptr(),
+            id.asRealmObjectIdT(),
+            callback
+        )
+    }
+
+    actual fun realm_app_user_apikey_provider_client_fetch_apikey(
+        app: RealmAppPointer,
+        user: RealmUserPointer,
+        id: ObjectId,
+        callback: AppCallback<ApiKeyWrapper>,
+    ) {
+        realmc.realm_app_user_apikey_provider_client_fetch_apikey(
+            app.cptr(),
+            user.cptr(),
+            id.asRealmObjectIdT(),
+            callback
+        )
+    }
+
+    actual fun realm_app_user_apikey_provider_client_fetch_apikeys(
+        app: RealmAppPointer,
+        user: RealmUserPointer,
+        callback: AppCallback<Array<ApiKeyWrapper>>,
+    ) {
+        realmc.realm_app_user_apikey_provider_client_fetch_apikeys(
+            app.cptr(),
+            user.cptr(),
+            callback
+        )
+    }
+
     private fun realm_value_t.asTimestamp(): Timestamp {
         if (this.type != realm_value_type_e.RLM_TYPE_TIMESTAMP) {
             error("Value is not of type Timestamp: $this.type")
@@ -1594,13 +1706,13 @@ actual object RealmInterop {
         return TimestampImpl(this.timestamp.seconds, this.timestamp.nanoseconds)
     }
 
-    private fun realm_value_t.asObjectId(): ObjectIdWrapper {
+    private fun realm_value_t.asObjectId(): ObjectId {
         if (this.type != realm_value_type_e.RLM_TYPE_OBJECT_ID) {
             error("Value is not of type ObjectId: $this.type")
         }
         val byteArray = ByteArray(OBJECT_ID_BYTES_SIZE)
         this.object_id.bytes.mapIndexed { index, b -> byteArray[index] = b.toByte() }
-        return ObjectIdWrapperImpl(byteArray)
+        return ObjectId(byteArray)
     }
 
     private fun realm_value_t.asUUID(): UUIDWrapper {
@@ -1693,16 +1805,9 @@ private fun capiRealmValue(realmValue: RealmValue): realm_value_t {
                     nanoseconds = value.nanoSeconds
                 }
             }
-            is ObjectIdWrapper -> {
+            is ObjectId -> {
                 cvalue.type = realm_value_type_e.RLM_TYPE_OBJECT_ID
-                cvalue.object_id = realm_object_id_t().apply {
-                    val data = ShortArray(OBJECT_ID_BYTES_SIZE)
-                    @OptIn(ExperimentalUnsignedTypes::class)
-                    (0 until OBJECT_ID_BYTES_SIZE).map {
-                        data[it] = value.bytes[it].toShort()
-                    }
-                    bytes = data
-                }
+                cvalue.object_id = value.asRealmObjectIdT()
             }
             is RealmObjectInterop -> {
                 val nativePointer = value.objectPointer
@@ -1739,6 +1844,17 @@ private fun capiRealmValue(realmValue: RealmValue): realm_value_t {
     return cvalue
 }
 
+private fun ObjectId.asRealmObjectIdT(): realm_object_id_t {
+    return realm_object_id_t().apply {
+        val data = ShortArray(OBJECT_ID_BYTES_SIZE)
+        val objectIdBytes = this@asRealmObjectIdT.toByteArray()
+        (0 until OBJECT_ID_BYTES_SIZE).map {
+            data[it] = objectIdBytes[it].toShort()
+        }
+        bytes = data
+    }
+}
+
 private class JVMScheduler(dispatcher: CoroutineDispatcher) {
     val scope: CoroutineScope = CoroutineScope(dispatcher)
 
@@ -1753,7 +1869,3 @@ private class JVMScheduler(dispatcher: CoroutineDispatcher) {
         )
     }
 }
-
-// using https://developer.android.com/reference/java/lang/System#getProperties()
-private fun isAndroid(): Boolean =
-    System.getProperty("java.specification.vendor")?.contains("Android") ?: false