+++ packages/test-base/src/commonMain/kotlin/io/realm/kotlin/entities/Sample.kt
@@ -17,6 +17,7 @@
 
 package io.realm.kotlin.entities
 
+import io.realm.kotlin.ext.backlinks
 import io.realm.kotlin.ext.realmListOf
 import io.realm.kotlin.ext.realmSetOf
 import io.realm.kotlin.types.MutableRealmInt
@@ -26,6 +27,7 @@ import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.types.RealmSet
 import io.realm.kotlin.types.RealmUUID
+import org.mongodb.kbson.BsonObjectId
 
 @Suppress("MagicNumber")
 class Sample : RealmObject {
@@ -40,6 +42,7 @@ class Sample : RealmObject {
     var doubleField: Double = 1.19840122
     var timestampField: RealmInstant = RealmInstant.from(100, 1000)
     var objectIdField: ObjectId = ObjectId.from("507f1f77bcf86cd799439011")
+    var bsonObjectIdField: BsonObjectId = BsonObjectId("507f1f77bcf86cd799439011")
     var uuidField: RealmUUID = RealmUUID.from("46423f1b-ce3e-4a7e-812f-004cf9c42d76")
     var binaryField: ByteArray = byteArrayOf(42)
     var mutableRealmIntField: MutableRealmInt = MutableRealmInt.create(42)
@@ -55,6 +58,7 @@ class Sample : RealmObject {
     var nullableDoubleField: Double? = null
     var nullableTimestampField: RealmInstant? = null
     var nullableObjectIdField: ObjectId? = null
+    var nullableBsonObjectIdField: ObjectId? = null
     var nullableUUIDField: RealmUUID? = null
     var nullableBinaryField: ByteArray? = null
     var nullableMutableRealmIntField: MutableRealmInt? = null
@@ -71,6 +75,7 @@ class Sample : RealmObject {
     var doubleListField: RealmList<Double> = realmListOf()
     var timestampListField: RealmList<RealmInstant> = realmListOf()
     var objectIdListField: RealmList<ObjectId> = realmListOf()
+    var bsonObjectIdListField: RealmList<BsonObjectId> = realmListOf()
     var uuidListField: RealmList<RealmUUID> = realmListOf()
     var binaryListField: RealmList<ByteArray> = realmListOf()
     var objectListField: RealmList<Sample> = realmListOf()
@@ -86,6 +91,7 @@ class Sample : RealmObject {
     var nullableDoubleListField: RealmList<Double?> = realmListOf()
     var nullableTimestampListField: RealmList<RealmInstant?> = realmListOf()
     var nullableObjectIdListField: RealmList<ObjectId?> = realmListOf()
+    var nullableBsonObjectIdListField: RealmList<BsonObjectId?> = realmListOf()
     var nullableUUIDListField: RealmList<RealmUUID?> = realmListOf()
     var nullableBinaryListField: RealmList<ByteArray?> = realmListOf()
 
@@ -100,6 +106,7 @@ class Sample : RealmObject {
     var doubleSetField: RealmSet<Double> = realmSetOf()
     var timestampSetField: RealmSet<RealmInstant> = realmSetOf()
     var objectIdSetField: RealmSet<ObjectId> = realmSetOf()
+    var bsonObjectIdSetField: RealmSet<BsonObjectId> = realmSetOf()
     var uuidSetField: RealmSet<RealmUUID> = realmSetOf()
     var binarySetField: RealmSet<ByteArray> = realmSetOf()
     var objectSetField: RealmSet<Sample> = realmSetOf()
@@ -115,8 +122,13 @@ class Sample : RealmObject {
     var nullableDoubleSetField: RealmSet<Double?> = realmSetOf()
     var nullableTimestampSetField: RealmSet<RealmInstant?> = realmSetOf()
     var nullableObjectIdSetField: RealmSet<ObjectId?> = realmSetOf()
+    var nullableBsonObjectIdSetField: RealmSet<BsonObjectId?> = realmSetOf()
     var nullableUUIDSetField: RealmSet<RealmUUID?> = realmSetOf()
     var nullableBinarySetField: RealmSet<ByteArray?> = realmSetOf()
+
+    val objectBacklinks by backlinks(Sample::nullableObject)
+    val listBacklinks by backlinks(Sample::objectListField)
+    val setBacklinks by backlinks(Sample::objectSetField)
 
     // For verification that references inside class is also using our modified accessors and are
     // not optimized to use the backing field directly.