+++ packages/test-base/src/commonMain/kotlin/io/realm/kotlin/entities/schema/SchemaVariations.kt
@@ -25,6 +25,7 @@ import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.types.RealmUUID
 import io.realm.kotlin.types.annotations.Index
 import io.realm.kotlin.types.annotations.PrimaryKey
+import org.mongodb.kbson.BsonObjectId
 
 /**
  * Class used for testing of the schema API; thus, doesn't exhaust modeling features but provides
@@ -46,6 +47,7 @@ class SchemaVariations : RealmObject {
     var string: String = "Realm"
     var date: RealmInstant = RealmInstant.from(0, 0)
     var objectId: ObjectId = ObjectId.create()
+    var bsonObjectId: BsonObjectId = BsonObjectId()
     var uuid: RealmUUID = RealmUUID.random()
     var binary: ByteArray = byteArrayOf(22, 66)
 
@@ -65,6 +67,7 @@ class SchemaVariations : RealmObject {
     var stringList: RealmList<String> = realmListOf()
     var dateList: RealmList<RealmInstant> = realmListOf()
     var objectIdList: RealmList<ObjectId> = realmListOf()
+    var bsonObjectIdList: RealmList<BsonObjectId> = realmListOf()
     var uuidList: RealmList<RealmUUID> = realmListOf()
     var binaryList: RealmList<ByteArray> = realmListOf()