+++ CHANGELOG.md
@@ -1,13 +1,23 @@
-## 1.4.1 (YYYY-MM-DD)
+## 1.5.0 (2022-11-11)
 
 ### Breaking Changes
 * None.
 
 ### Enhancements
-* None.
+* Fixed error when using Realm object as query argument. Issue[#1098](https://github.com/realm/realm-kotlin/issues/1098)
+* Realm will now use `System.loadLibrary()` first when loading native code on JVM, adding support for 3rd party JVM installers. If this fails, it will fallback to the current method of extracting and loading the native library from the JAR file. (Issue [#1105](https://github.com/realm/realm-kotlin/issues/1105)). 
+* Added support for in-memory Realms.
+* [Sync] Added support for API key authentication. (Issue [#432](https://github.com/realm/realm-kotlin/issues/432))
+* Added support for reverse relationships through the `linkingObjects` delegate. See the function documentation for more details. (Issue [#1021](https://github.com/realm/realm-kotlin/pull/1021))
+* Added support for `BsonObjectId` and its typealias `org.mongodb.kbson.ObjectId` as a replacement for `ObjectId`. `io.realm.kotlin.types.ObjectId` is still functional but has been marked as deprecated.
+* [Sync] Added support for `BsonObjectId` as partition value.
+* [Sync] Exposed `configuration` and `user` on `SyncSession`. (Issue [#431](https://github.com/realm/realm-kotlin/issues/431))
+* [Sync] Added support for encrypting the user metadata used by Sync. (Issue [#413](https://github.com/realm/realm-kotlin/issues/413))
 
 ### Fixed
-* Fixed error when using Realm object as query argument. Issue[#1098](https://github.com/realm/realm-kotlin/issues/1098)
+* Close underlying realm if it is no longer referenced by any Kotlin object. (Issue [#671](https://github.com/realm/realm-kotlin/issues/671))
+* Fixes crash during migration if Proguard was enabled. (Issue [#1106](https://github.com/realm/realm-kotlin/issues/1106))
+* Adds missing Proguard rules for Embedded objects. (Issue [#1106](https://github.com/realm/realm-kotlin/issues/1107))
 
 ### Compatibility
 * This release is compatible with the following Kotlin releases:
@@ -21,7 +31,10 @@
 * Minimum Android SDK: 16.
 
 ### Internal
-* None
+* Added dependency Kbson 0.1.0.
+* Updated to use hierarchical multi platform project structure.
+* Updated to Realm Core 12.11.0, commit 3d5ff9b5e47c5664c4c5611cdfd22fd15e451b55.
+* Updated to Detekt 1.22.0-RC2.
 
 
 ## 1.4.0 (2022-10-17)